fix(code): split Transaction.swift, MARK-up AnalysisStore, inject monthEnd
#654: Split `Domain/Models/Transaction.swift` (386 lines, three primary
types) into three files matching the one-primary-type-per-file
convention:
- `Transaction.swift` keeps the `Transaction` struct + structure-query
  computed properties (~85 lines).
- `TransactionFilter.swift` holds the filter struct + `hasActiveFilters`
  extension.
- `TransactionPage.swift` holds the page struct, the running-balance
  algorithm (`withRunningBalances` + helpers), and the file-private
  logger that the running-balance code uses.

#657: Add `// MARK: -` section dividers to `AnalysisStore.swift`
(State, Filters, Dependencies, Lifecycle, Data Loading, Aggregation,
Date Utilities) — file is 321 lines and the threshold is 300.

#658: Replace the `monthEnd` stored-property inline initialiser with a
defaulted init parameter
(`monthEnd: Int = Calendar.current.component(.day, from: Date())`).
The property is now a plain stored `Int` set exclusively in `init`,
unifying the initialisation path and giving tests a clean override
seam — they no longer depend on wall-clock at construction time.
Moves the property up next to the other observed state with a
clarifying doc comment.

Fixes #654
Fixes #657
Fixes #658
